Rename dci to lmi
