+++ rs/src/fql/backend.rs
@@ -1,6 +1,6 @@
 pub mod sqlexec;
 pub mod dri; /* Dynamically-typed Row-based Interpreter */
-pub mod dci; /* Dynamically-typed Column-based Interpreter */
+pub mod lmi; /* Late Materialize Interpreter */
 pub mod sci; /* Statically-typed Column-based Interpreter */
 
 use crate::data::DB_FILENAME;
